fix(repo): address beta program review feedback
